requirements.txt
pandas
openpyxl        # pandas 读取 xlsx 需要它
sqlalchemy
pymysql         # MySQL 的驱动
python-dotenv   # 读取 .env
streamlit
cryptography
plotly==6.4.0
scipy
pytest
watchdog